requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.4
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0

mysql-connector-python~=9.1.0
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
pytest~=8.3.3
PyJWT~=2.9.0
passlib~=1.7.4